api/requirements.txt
numpy>=1.26.0
opencv-python>=4.8.0.76
pygame>=2.5.0
pillow>=10.0.0
PyYAML>=6.0.2
SQLAlchemy>=2.0.30
alembic>=1.13.1
psycopg[binary]>=3.1.19

sounddevice>=0.4.6
soundfile>=0.12.1
SpeechRecognition>=3.10.0
pocketsphinx>=5.0.0
PyAudio>=0.2.14
gTTS>=2.4.0
playsound>=1.2.2

PyOpenGL>=3.1.7

openai>=1.35.4,<2.0.0
python-dotenv>=1.0.0

# FastAPI関連
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.9
websockets>=12.0
